% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listener_core.R
\name{add_listener}
\alias{add_listener}
\title{Internal function to add listeners}
\usage{
add_listener(jug, func, event)
}
\arguments{
\item{jug}{the jug instance}

\item{func}{the function to bind}

\item{event}{the event to bind to}
}
\description{
Internal function to add listeners
}
