% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{add_middleware}
\alias{add_middleware}
\title{Internal function to add middleware}
\usage{
add_middleware(jug, func, path = NULL, method = NULL)
}
\arguments{
\item{jug}{the jug instance}

\item{func}{the function to bind}

\item{path}{the path to bind to}

\item{method}{the method to bind to}
}
\description{
Internal function to add middleware
}

