% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{add_path}
\alias{add_path}
\title{Add function to interface definition}
\usage{
add_path(obj, path, func, res_content_type = "application/json")
}
\arguments{
\item{obj}{the interface definition}

\item{func}{the function to add}
}
\description{
Add function to interface definition
}

