% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware_collector.R
\name{collector}
\alias{collector}
\title{Collector function for middlewares}
\usage{
collector()
}
\description{
Basically constructs an empty jug instances that will be merged into the primary one later on
}

