% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{decorate}
\alias{decorate}
\title{Convenience function to decorate existing functions}
\usage{
decorate(func, content_type = "text/html", strict_params = FALSE)
}
\arguments{
\item{func}{the function to decorate}

\item{content_type}{the content type to set on the response object}

\item{strict_params}{if TRUE and not all parameters needed for func}
}
\description{
Convenience function to decorate existing functions
}
