% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\docType{data}
\name{Error}
\alias{Error}
\title{Response class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0758d0e4> 
 - attr(*, "name")= chr "Error_generator"
}}
\usage{
Error
}
\description{
Response class
}
\keyword{datasets}

