% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{get.Jug}
\alias{get.Jug}
\title{Function to add GET-binding middleware}
\usage{
\method{get}{Jug}(jug, path, ...)
}
\arguments{
\item{jug}{the jug object}

\item{path}{the path to bind to}

\item{...}{functions (order matters) to bind to the path (will receive the params \code{req}, \code{res} and \code{err})}
}
\description{
Function to add GET-binding middleware
}

