% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request.R
\name{get_passed_params}
\alias{get_passed_params}
\title{Helper function to deparse query params}
\usage{
get_passed_params(query_string)
}
\arguments{
\item{query_string}{the request object}
}
\description{
Helper function to deparse query params
}

