% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_collector.R
\name{include}
\alias{include}
\title{Include elsewhere constructed middleware with primary jug instance}
\usage{
include(jug, collector, source_file = NULL)
}
\arguments{
\item{jug}{the primary jug instance}

\item{collector}{the variable containing the external middleware}

\item{source_file}{provide (relative) path if collector variable is located in another .R file}
}
\description{
Include elsewhere constructed middleware with primary jug instance
}

