% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listener_core.R
\docType{data}
\name{Listener}
\alias{Listener}
\title{Middleware class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Listener
}
\description{
Middleware class
}
\keyword{datasets}
