% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listener_logger.R
\name{logger}
\alias{logger}
\title{Basic logger}
\usage{
logger(jug, threshold = futile.logger::DEBUG, log_file = NULL,
  console = TRUE)
}
\arguments{
\item{jug}{a jug instance}

\item{threshold}{the threhold as from what type of message to activate the logger (check the documentation of \pkg{futile.logger} for possibilities)}

\item{log_file}{if given, the file to which the log is written}

\item{console}{if TRUE will (also) print to console}
}
\description{
This logger uses futile.logger to implement logging functionality
}
