% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{match_path}
\alias{match_path}
\title{Helper function to extract regex named capture groups from string}
\usage{
match_path(pattern, path, ...)
}
\arguments{
\item{pattern}{regex pattern}

\item{path}{the path to test the regex pattern against}

\item{...}{arguments passed to \code{regexpr}}
}
\description{
adapted from: http://oddhypothesis.blogspot.be/2012/05/regex-named-capture-in-r.html
}
