% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\docType{data}
\name{Middleware}
\alias{Middleware}
\title{Middleware class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x06549e50> 
 - attr(*, "name")= chr "Middleware_generator"
}}
\usage{
Middleware
}
\description{
Middleware class
}
\keyword{datasets}

