% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware_core.R
\docType{data}
\name{MiddlewareHandler}
\alias{MiddlewareHandler}
\title{MiddlewareHandler R6 class definition}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MiddlewareHandler
}
\description{
MiddlewareHandler R6 class definition
}
\keyword{datasets}
