% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listener_core.R
\name{on}
\alias{on}
\title{Function to add an event listener}
\usage{
on(jug, event, func, ...)
}
\arguments{
\item{jug}{the jug instance}

\item{event}{the event to bind to (one of \code{start}, \code{finish} or \code{error})}

\item{func}{the function to pass the \code{req}, \code{res}, \code{err} and (if \code{error} event) the \code{err_msg} to}

\item{...}{extra arguments to pass to the listener function}
}
\description{
Function to add an event listener
}
\seealso{
\code{\link{post}}, \code{\link{put}}, \code{\link{delete}}, \code{\link{use}}, \code{\link{ws}}
}
