% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_params}
\alias{parse_params}
\title{Parse the params passed by the request}
\usage{
parse_params(body, query_string = NULL, content_type)
}
\arguments{
\item{content_type}{the mime type}

\item{env}{the rook req environment}
}
\description{
Parse the params passed by the request
}

