% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_post_data}
\alias{parse_post_data}
\title{Parse the post data}
\usage{
parse_post_data(env, content_type)
}
\arguments{
\item{env}{the rook req environment}

\item{content_type}{the mime type}
}
\description{
Parse the post data
}

