% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_response}
\alias{plot_response}
\title{Convenience function to return plots in a response (as a PNG image)}
\usage{
plot_response(res, plot_obj, ...)
}
\arguments{
\item{res}{the response instance}

\item{plot_obj}{the plot object}

\item{...}{parameters passed to the png device}
}
\description{
Convenience function to return plots in a response (as a PNG image)
}

