% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{post}
\alias{post}
\title{Function to add POST-binding middleware}
\usage{
post(jug, path, ...)
}
\arguments{
\item{jug}{the jug object}

\item{path}{the path to bind to}

\item{...}{functions (order matters) to bind to the path (will receive the params \code{req}, \code{res} and \code{err})}
}
\description{
Function to add POST-binding middleware
}

