% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_request.R
\name{process_test_request}
\alias{process_test_request}
\title{Initialize process of test request}
\usage{
process_test_request(jug, test_request)
}
\arguments{
\item{jug}{the jug instance}

\item{test_request}{the RawTestRequest instance}
}
\description{
Initialize process of test request
}

