% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{re_capture}
\alias{re_capture}
\title{Helper function to extract regex named capture groups from string}
\usage{
re_capture(pattern, string, ...)
}
\arguments{
\item{pattern}{regex pattern}

\item{string}{the string to test the regex pattern against}

\item{...}{arguments passed to \code{regexpr}}
}
\description{
source: http://oddhypothesis.blogspot.be/2012/05/regex-named-capture-in-r.html
}

