% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request.R
\docType{data}
\name{Request}
\alias{Request}
\title{Request class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0651327c> 
 - attr(*, "name")= chr "Request_generator"
}}
\usage{
Request
}
\description{
Request class
}
\keyword{datasets}

