% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jug.R
\docType{data}
\name{RequestHandler}
\alias{RequestHandler}
\title{RequestHandler R6 class definition}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RequestHandler
}
\description{
RequestHandler R6 class definition
}
\keyword{datasets}
