% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_collector.R
\name{require_middleware}
\alias{require_middleware}
\title{Join elsewhere constructed middleware with primary jug instance}
\usage{
require_middleware(jug, collector, file_to_source = NULL)
}
\arguments{
\item{jug}{the primary jug instance}

\item{collector}{the variable containing the external middleware}
}
\description{
Join elsewhere constructed middleware with primary jug instance
}

