% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/response.R
\docType{data}
\name{Response}
\alias{Response}
\title{Response class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x06dcecd4> 
 - attr(*, "name")= chr "Response_generator"
}}
\usage{
Response
}
\description{
Response class
}
\keyword{datasets}

