% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{Server}
\alias{Server}
\title{The Server class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x065ab244> 
 - attr(*, "name")= chr "Server_generator"
}}
\usage{
Server
}
\description{
The Server class
}
\keyword{datasets}

