% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_error.R
\name{simple_error_handler}
\alias{simple_error_handler}
\title{An error handler middleware}
\usage{
simple_error_handler(jug, path = NULL, to_cat_or_not_to_cat = "cat")
}
\arguments{
\item{jug}{the jug instance}

\item{to_cat_or_not_to_cat}{the argument says it all}

\item{the}{path to bind to, default = NULL (all paths)}
}
\description{
An error handler middleware
}

