% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware_error.R
\name{simple_error_handler_json}
\alias{simple_error_handler_json}
\title{An error handler middleware which returns a error description in JSON format}
\usage{
simple_error_handler_json(jug, path = NULL)
}
\arguments{
\item{jug}{the jug instance}

\item{path}{the path to bind to, default = NULL (all paths)}
}
\description{
An error handler middleware which returns a error description in JSON format
}
\seealso{
simple_error_handler
}

