% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{start_interface}
\alias{start_interface}
\title{start the interface}
\usage{
start_interface(obj, host = "127.0.0.1", port = 8083,
  verbose = getOption("verbose"))
}
\arguments{
\item{obj}{the interface object}
}
\description{
start the interface
}

