% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{use}
\alias{use}
\title{Function to add request method insensitive middleware}
\usage{
use(jug, path, ..., method = NULL)
}
\arguments{
\item{jug}{the jug object}

\item{path}{the path to bind to}

\item{...}{functions (order matters) to bind to the path (will receive the params \code{req}, \code{res} and \code{err})}

\item{method}{the method to bind to, defauts to \code{NULL}}
}
\description{
Function to add request method insensitive middleware
}
\seealso{
\code{\link{post}}, \code{\link{put}}, \code{\link{delete}}, \code{\link{get}}, \code{\link{ws}}
}

