% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{web_plot}
\alias{web_plot}
\title{Convenience function to return plots in a response (as a PNG image)}
\usage{
web_plot(plot_obj, res, ...)
}
\arguments{
\item{plot_obj}{the plot object}

\item{res}{the response instance}

\item{...}{parameters passed to the png device}
}
\description{
Convenience function to return plots in a response (as a PNG image)
}

