% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{add_middleware}
\alias{add_middleware}
\title{Internal function to add middleware}
\usage{
add_middleware(jug, func, path = NULL, method = NULL, websocket = FALSE)
}
\arguments{
\item{jug}{the jug instance}

\item{func}{the function to bind}

\item{path}{the path to bind to}

\item{method}{the method to bind to}

\item{websocket}{should the middleware bind to the websocket protocol}
}
\description{
Internal function to add middleware
}

