% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decorators.R
\name{decorate}
\alias{decorate}
\title{Convenience function to decorate existing functions}
\usage{
decorate(func, content_type = "text/html")
}
\arguments{
\item{func}{the function to decorate}
}
\description{
Convenience function to decorate existing functions
}

