% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{Convenience function to decorate existing functions}
\usage{
decorate(func)
}
\arguments{
\item{func}{the function to decorate}
}
\description{
Convenience function to decorate existing functions
}

