% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{get.Jug}
\alias{get.Jug}
\title{Function to add GET-binding middleware}
\usage{
\method{get}{Jug}(jug, path, func)
}
\arguments{
\item{jug}{the jug object}

\item{path}{the path to bind to}

\item{func}{the function to bind to the path (will receive the params \code{req} and \code{res})}
}
\description{
Function to add GET-binding middleware
}

