% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{get}
\alias{get}
\title{Set up generic method (otherwise base::get is masked)}
\usage{
get(object, ...)
}
\arguments{
\item{object}{the object to pass to get}

\item{...}{other arguments passed to get}
}
\description{
Set up generic method (otherwise base::get is masked)
}

