% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{get}
\alias{get}
\title{Function to add GET-binding middleware}
\usage{
get(object, ...)
}
\arguments{
\item{object}{the jug object}

\item{...}{functions (order matters) to bind to the path (will receive the params \code{req}, \code{res} and \code{err})}
}
\description{
Function to add GET-binding middleware
}

