% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listener_logger.R
\name{logger}
\alias{logger}
\title{Basic logging middleware}
\usage{
logger(jug, threshold = futile.logger::DEBUG, log_file = NULL,
  console = TRUE)
}
\arguments{
\item{log_file}{the text file to which the log will be appended}

\item{log_msg_func}{the function which returns the content of the log message, expects a function that allows to pass a \code{req}, \code{res} and \code{err} argument}
}
\description{
Basic logging middleware
}
