% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\docType{data}
\name{MiddlewareHandler}
\alias{MiddlewareHandler}
\title{MiddlewareHandler R6 class definition}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x06d751a4> 
 - attr(*, "name")= chr "MiddlewareHandler_generator"
}}
\usage{
MiddlewareHandler
}
\description{
MiddlewareHandler R6 class definition
}
\keyword{datasets}

