% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_params}
\alias{parse_params}
\title{Parse the params passed by the request}
\usage{
parse_params(req, body, query_string, content_type)
}
\arguments{
\item{req}{the rook req environment}

\item{body}{the parsed body}

\item{query_string}{a query string}

\item{content_type}{the mime type}
}
\description{
Parse the params passed by the request
}
