% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{post}
\alias{post}
\title{Function to add POST middleware}
\usage{
post(jug, path, func)
}
\arguments{
\item{jug}{the jug object}

\item{path}{the path to bind to}

\item{func}{the function to bind to the path (will receive the params \code{req} and \code{res})}
}
\description{
Function to add POST middleware
}

