% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_request.R
\docType{data}
\name{RawTestRequest}
\alias{RawTestRequest}
\title{Generate request for testing purposes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RawTestRequest
}
\description{
solely used for testing purposes
}
\keyword{datasets}
