% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_request.R
\docType{data}
\name{RawTestRequest}
\alias{RawTestRequest}
\title{Generate request for testing purposes}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x062336f0> 
 - attr(*, "name")= chr "RawTestRequest_generator"
}}
\usage{
RawTestRequest
}
\description{
solely used for testing purposes
}
\keyword{datasets}

