% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jug.R
\name{serve_it}
\alias{serve_it}
\title{Start serving the jug instance}
\usage{
serve_it(jug, host = "127.0.0.1", port = 8080, daemonized = FALSE)
}
\arguments{
\item{jug}{the jug instance}

\item{host}{the host address}

\item{port}{the port to host on}

\item{daemonized}{whether or not to start a daemonized server (experimental)}
}
\description{
Start serving the jug instance
}

