% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_static.R
\name{serve_static_files}
\alias{serve_static_files}
\title{Middleware to serve static files}
\usage{
serve_static_files(jug, path = NULL, root_path = getwd())
}
\arguments{
\item{jug}{the jug instance}

\item{root_path}{}

\item{the}{path to bind to, default = NULL (all paths)}
}
\description{
Binds to get requests that aren't handled by specified paths.
Currently only works/tested for text files
}

