% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware_static.R
\name{serve_static_files}
\alias{serve_static_files}
\title{Middleware to serve static files}
\usage{
serve_static_files(jug, path = NULL, root_path = getwd())
}
\arguments{
\item{jug}{the jug instance}

\item{path}{the path to bind to, default = NULL (all paths)}

\item{root_path}{the file path to set as root for the file server}
}
\description{
Binds to get requests that aren't handled by specified paths.
Should support all filetypes; returns image and octet-stream types as a raw string.
}
