% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/middleware_core.R
\name{use}
\alias{use}
\title{Function to add request method insensitive middleware}
\usage{
use(jug, path, ...)
}
\arguments{
\item{jug}{the jug object}

\item{path}{the path to bind to}

\item{func}{the function to bind to the path (will receive the params \code{req}, \code{res} and \code{err})}
}
\description{
Function to add request method insensitive middleware
}

